Fix bug where webext fetch message listener kills all others.
Returning a Promise from an onMessage event listener means the resolved value will be sent back as a response.

This means if our message condition isn't met so we don't return a value, the async function will return a Promise that resolves to null.

Therefore, this handler will send a null value back to all messages, and any handlers that haven't yet fired which may mach the message condition are no longer able to send a reply.

This also causes the tab to crash in Firefox. Possibly because Firefox is adding it's own native handlers which no longer communicate back to the page...?